PODB-438: Delete request logger class
