+++ src/LtiServiceConnector.php
@@ -6,7 +6,6 @@ use Firebase\JWT\JWT;
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\ClientException;
 use GuzzleHttp\Psr7\Response;
-use Packback\Lti1p3\Helpers\RequestLogger;
 use Packback\Lti1p3\Interfaces\ICache;
 use Packback\Lti1p3\Interfaces\ILtiRegistration;
 use Packback\Lti1p3\Interfaces\ILtiServiceConnector;
@@ -19,19 +18,28 @@ class LtiServiceConnector implements ILtiServiceConnector
     public const METHOD_GET = 'GET';
     public const METHOD_POST = 'POST';
 
+    // Supported request types which map to an error log message
+    public const UNSUPPORTED_REQUEST = 0;
+    public const SYNC_GRADE_REQUEST = 1;
+    public const CREATE_LINEITEM_REQUEST = 2;
+
     private $cache;
     private $client;
-    private $requestLogger;
     private $debuggingMode = false;
+    private $errorMessages;
 
     public function __construct(
         ICache $cache,
-        Client $client,
-        RequestLogger $requestLogger
+        Client $client
     ) {
         $this->cache = $cache;
         $this->client = $client;
-        $this->requestLogger = $requestLogger;
+
+        $this->errorMessages = [
+            static::UNSUPPORTED_REQUEST => 'Logging request data: ',
+            static::SYNC_GRADE_REQUEST => 'Syncing grade for this lti_user_id: ',
+            static::CREATE_LINEITEM_REQUEST => 'Creating lineitem for this lti_user_id: ',
+        ];
     }
 
     public function setDebuggingMode(bool $enable): void
@@ -112,7 +120,7 @@ class LtiServiceConnector implements ILtiServiceConnector
         // Set $requestType here, since static properties cannot be evaluated
         // as parameters
         if (!isset($requestType)) {
-            $requestType = RequestLogger::UNSUPPORTED_REQUEST;
+            $requestType = self::UNSUPPORTED_REQUEST;
         }
 
         $request->setAccessToken($this->getAccessToken($registration, $scopes));
@@ -140,7 +148,7 @@ class LtiServiceConnector implements ILtiServiceConnector
         $responseBody = $this->getResponseBody($response);
 
         if ($this->debuggingMode) {
-            $this->requestLogger->logRequest(
+            $this->logRequest(
                 $requestType,
                 $request,
                 $responseHeaders,
@@ -181,6 +189,37 @@ class LtiServiceConnector implements ILtiServiceConnector
         }
 
         return $results;
+    }
+
+    public function logRequest(
+        int $requestType,
+        IServiceRequest $request,
+        array $responseHeaders,
+        ?array $responseBody
+    ): void {
+        $requestBody = $request->getPayload()['body'];
+
+        $contextArray = [
+            'request_method' => $request->getMethod(),
+            'request_url' => $request->getUrl(),
+            'request_body' => $requestBody,
+            'response_headers' => $responseHeaders,
+            'response_body' => json_encode($responseBody),
+        ];
+
+        $this->errorLog(json_decode($requestBody)->userId, $contextArray);
+    }
+
+    /**
+     * A wrapper for the PHP error_log function to facilitate testing.
+     */
+    public function errorLog(
+        string $userId,
+        array $contextArray
+    ): void {
+        $logMsg = $this->errorMessages[$requestType];
+
+        error_log($logMsg.$userId.' '.print_r($contextArray, true));
     }
 
     private function getAccessTokenCacheKey(ILtiRegistration $registration, array $scopes)
